requirements.txt
Django==4.2.1
dj-database-url==2.0.0
django-cryptography==1.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
apiview==1.3.25
cryptography==41.0.1
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pyotp==2.8.0
python-dotenv==1.0.0
sqlparse==0.4.4
twilio==8.2.2
whitenoise==6.4.0
